api/admin_dashboard_endpoints.py
import sqlite3
"""
Admin Dashboard Enhanced API Endpoints
Provides comprehensive dashboard analytics and system health metrics
"""

from fastapi import APIRouter, HTTPException, Depends
from pydantic import BaseModel
from typing import List, Dict, Any, Optional
from datetime import datetime, timedelta
import json
import time
import asyncio
from sqlalchemy import text

from modules.core.logging import LogManager
from modules.core.auth_dependency import get_admin_user as require_admin, get_current_user
from modules.core.config import Config
from modules.core.db import DBManager

# Initialize components
logger = LogManager.setup_logging(__name__)
db_manager = DBManager()

router = APIRouter()

# Pydantic models
class SystemHealthStatus(BaseModel):
    overall: str  # healthy, warning, error
    api: str
    database: str
    documents: str
    details: Dict[str, Any]

class DashboardStatistics(BaseModel):
    active_users: int
    total_users: int
    documents_processed_today: int
    documents_processed_week: int
    documents_failed_today: int
    avg_response_time_ms: float
    total_storage_bytes: int
    user_trend_percent: float
    document_trend_percent: float
    response_trend_percent: float

class RAGMetrics(BaseModel):
    total_queries: int
    queries_today: int
    accuracy_percent: float
    cache_hit_rate: float
    avg_context_size: int
    avg_retrieval_time_ms: float
    total_embeddings: int
    query_trend_percent: float

class QueueStatistics(BaseModel):
    queued: int
    processing: int
    completed_today: int
    failed_today: int
    paused: bool
    avg_processing_time_ms: float

class RecentActivity(BaseModel):
    id: int
    type: str  # document, user, system, error
    description: str
    timestamp: datetime
    metadata: Dict[str, Any]

class DashboardSummary(BaseModel):
    health: SystemHealthStatus
    statistics: DashboardStatistics
    rag_metrics: RAGMetrics
    queue_stats: QueueStatistics
    recent_activities: List[RecentActivity]

# Helper functions
async def check_api_health() -> str:
    """Check API health status"""
    try:
        # Simple health check - could be expanded
        return "healthy"
    except Exception as e:
        logger.error(f"Error: {e}")
        raise HTTPException(status_code=500, detail=str(e))
    

@router.post("/queue/pause")
async def pause_queue(user: Dict[str, Any] = Depends(require_admin)):
    """Pause document processing queue"""
    try:
        from modules.background.job_manager import BackgroundJobManager
        job_manager = BackgroundJobManager()
        
        if hasattr(job_manager, 'pause_queue'):
            success = job_manager.pause_queue()
            return {"success": success}
        
        return {"success": True}  # Mock success
    except Exception as e:
        logger.error(f"Error: {e}")
        raise HTTPException(status_code=500, detail=str(e))
    

@router.post("/actions/optimize-database")
async def optimize_database(user: Dict[str, Any] = Depends(require_admin)):
    """Trigger database optimization"""
    try:
        # Start optimization job
        from modules.background.job_manager import BackgroundJobManager
        job_manager = BackgroundJobManager()
        
        if hasattr(job_manager, 'create_job'):
            job_id = job_manager.create_job(
                name="Database Optimization",
                type="optimize_db",
                params={}
            )
            return {"success": True, "job_id": job_id}
        
        return {"success": True, "message": "Optimization started"}
    except Exception as e:
        logger.error(f"Error: {e}")
        raise HTTPException(status_code=500, detail=str(e))


@router.get("/summary")
async def get_dashboard_summary(user: Dict[str, Any] = Depends(require_admin)):
    """Get comprehensive dashboard summary"""
    try:
        with db_manager.get_session() as session:
            # Get active users stats
            now = datetime.now()
            today_start = now.replace(hour=0, minute=0, second=0, microsecond=0)
            week_start = now - timedelta(days=7)
            month_start = now - timedelta(days=30)
            
            active_today = session.execute(text("""
                SELECT COUNT(DISTINCT user_id) 
                FROM chat_sessions 
                WHERE created_at >= :start
            """), {"start": today_start}).scalar() or 0
            
            active_week = session.execute(text("""
                SELECT COUNT(DISTINCT user_id) 
                FROM chat_sessions 
                WHERE created_at >= :start
            """), {"start": week_start}).scalar() or 0
            
            active_month = session.execute(text("""
                SELECT COUNT(DISTINCT user_id) 
                FROM chat_sessions 
                WHERE created_at >= :start
            """), {"start": month_start}).scalar() or 0
            
            # Get document stats (using feedback as proxy for document processing)
            total_feedback = session.execute(text("""
                SELECT COUNT(*) FROM feedback
            """)).scalar() or 0
            
            # For demo purposes, simulate queued and failed
            queued = 5  # Mock value
            failed = 2  # Mock value
            
            # Get RAG stats
            total_messages = session.execute(text("""
                SELECT COUNT(*) FROM chat_messages
            """)).scalar() or 0
            
            # Calculate average response time (mock for now)
            avg_response_time = 180.5  # milliseconds
            
            # Mock RAG accuracy and cache hit rate
            avg_accuracy = 92.5
            cache_hit_rate = 78.3
            
            # System health check
            system_health = {
                "api": "healthy",
                "database": "healthy",
                "rag": "healthy",
                "cache": "healthy",
                "documents": "healthy"
            }
            
            # Check if database is responding
            try:
                session.execute(text("SELECT 1")).scalar()
            except:
                system_health["database"] = "error"
            
            return {
                "activeUsers": {
                    "today": active_today,
                    "week": active_week,
                    "month": active_month
                },
                "documentStats": {
                    "processed": total_feedback,
                    "queued": queued,
                    "failed": failed
                },
                "ragStats": {
                    "totalQueries": total_messages,
                    "avgAccuracy": avg_accuracy,
                    "cacheHitRate": cache_hit_rate,
                    "avgResponseTime": avg_response_time
                },
                "responseTimeAvg": avg_response_time,
                "systemHealth": system_health
            }
            
    except Exception as e:
        logger.error(f"Error getting dashboard summary: {e}")
        raise HTTPException(status_code=500, detail=str(e))
    
